pyproject.toml
[project]
name = "axonforge"
version = "0.1.0"
description = "A node-based computational framework with a PySide6 visual editor for building and simulating neural-inspired systems"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "pyside6>=6.7.0",
    "python-mnist>=0.7",
]

[tool.setuptools.packages.find]
include = ["axonforge*", "axonforge_qt*"]

[project.scripts]
axonforge = "axonforge.cli:main"
